Moved MockFileShimBehavior and MockDirectoryShimBehavior to ShimBehaviors folder.
