+++ MissingFeatures.Tests/DirectoryHelper/ShimBehaviors/MockFileShimBehavior.cs
@@ -1,0 +1,36 @@
+﻿namespace MissingFeatures.Tests.DirectoryHelper.ShimBehaviors
+{
+    using System;
+    using System.IO.Abstractions.TestingHelpers;
+    using System.Linq;
+    using System.Reflection;
+
+    using Microsoft.QualityTools.Testing.Fakes.Shims;
+
+    public class MockFileShimBehavior : IShimBehavior
+    {
+        private readonly MockFile mockFile;
+
+        public MockFileShimBehavior(IMockFileDataAccessor fileDataAccessor)
+        {
+            this.mockFile = new MockFile(fileDataAccessor);
+        }
+
+        public bool TryGetShimMethod(MethodBase method, out Delegate shim)
+        {
+            // TODO: think what happens when a static method is called
+            if (method is MethodInfo)
+            {
+                var methodParameterTypes = method.GetParameters().Select(p => p.ParameterType).ToArray();
+                var methodInfo = typeof(MockFile).GetMethod(method.Name, methodParameterTypes);
+
+                shim = methodInfo.CreateDelegate(this.mockFile);
+                return true;
+            }
+
+            // No shim method for constuctors.
+            shim = null;
+            return false;
+        }
+    }
+}
